Reorganize families and results
