+++ src/solver.h
@@ -2,46 +2,14 @@
 
 #include <RcppArmadillo.h>
 #include <memory>
-#include "families.h"
+#include "families/families.h"
 #include "utils.h"
 #include "prox.h"
 #include "infeasibility.h"
+#include "results.h"
 
 using namespace arma;
 using namespace Rcpp;
-
-struct Results {
-  rowvec intercept;
-  mat beta;
-  uword passes;
-  std::vector<double> primals;
-  std::vector<double> duals;
-  std::vector<double> infeasibilities;
-  std::vector<double> time;
-  std::vector<unsigned> line_searches;
-  double deviance;
-
-  Results() {}
-
-  Results(rowvec intercept,
-          mat beta,
-          uword passes,
-          std::vector<double> primals,
-          std::vector<double> duals,
-          std::vector<double> infeasibilities,
-          std::vector<double> time,
-          std::vector<unsigned> line_searches,
-          double deviance)
-          : intercept(intercept),
-            beta(beta),
-            passes(passes),
-            primals(primals),
-            duals(duals),
-            infeasibilities(infeasibilities),
-            time(time),
-            line_searches(line_searches),
-            deviance(deviance) {}
-};
 
 class Solver {
 private:
